Revert "fix: RND54-checkbox"
This reverts commit 89807cb9adfda435832e5df52166c4fc21bdc4d5.

+++ Frontend/src/pages/admin/ServiceManagement.jsx
@@ -106,6 +106,7 @@ const ServiceManagement = () => {
 
     // --- 5. DELETE SERVICE ---
     const handleDelete = async (id) => {
+        if (window.confirm("Are you sure you want to delete this service?")) {
             try {
                 const token = localStorage.getItem('token');
                 await axios.delete(`http://localhost:8081/api/admin/services/${id}`, {
@@ -117,6 +118,7 @@ const ServiceManagement = () => {
                 console.error("Error deleting service:", error);
                 toast.error("Failed to delete this service.");
             }
+        }
     };
 
     return (
